refactor(analytics): remove AnalyticsServiceBase and obsolete static config props
Remove `AnalyticsServiceBase` by moving the required code to `AnalyticsService`.
Remove obsolete static config props.
