+++ lib/common/definitions/config.d.ts
@@ -6,8 +6,6 @@ declare module Config {
 		USER_AGENT_NAME: string;
 		CLIENT_NAME_ALIAS?: string;
 		FULL_CLIENT_NAME?: string;
-		ANALYTICS_API_KEY: string;
-		ANALYTICS_EXCEPTIONS_API_KEY: string;
 		ANALYTICS_INSTALLATION_ID_SETTING_NAME: string;
 		TRACK_FEATURE_USAGE_SETTING_NAME: string;
 		ERROR_REPORT_SETTING_NAME: string;
